+++ src/RePraxis/QueryResult.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using System.Text;
 
 namespace RePraxis
 {
@@ -29,6 +30,33 @@ namespace RePraxis
 		{
 			Success = success;
 			Bindings = new Dictionary<string, string>[0];
+		}
+
+		/// <summary>
+		/// Return a prettified string explaining the result
+		/// </summary>
+		/// <returns></returns>
+		public string ToPrettyString()
+		{
+			StringBuilder sb = new StringBuilder();
+
+			sb.AppendLine( $"Result Success: {Success}" );
+			sb.AppendLine( "Bindings: [" );
+
+			for ( int i = 0; i < Bindings.Length; i++ )
+			{
+				sb.Append( $"[{i}] " );
+				sb.Append( "{" );
+				sb.Append(
+				string.Join(
+					", ", Bindings[i].Select( b => b.Key + "=" + b.Value ).ToArray()
+					)
+				);
+				sb.Append( "}\n" );
+			}
+			sb.AppendLine( "]" );
+
+			return sb.ToString();
 		}
 	}
 }