+++ tests/RePraxis.Tests/TestRePraxis.cs
@@ -6,7 +6,7 @@ public class Tests
 	/// Test that create, retrieval, updating, and deletion (CRUD)
 	/// </summary>
 	[Test]
-	public void TestDatabaseCRUD()
+	public void TestInsertSentence()
 	{
 		var db = new RePraxisDatabase();
 
@@ -18,14 +18,39 @@ public class Tests
 		Assert.That( db.Assert( "A.relationships.B.reputation!19" ), Is.EqualTo( false ) );
 		Assert.That( db.Assert( "A.relationships.B.type" ), Is.EqualTo( true ) );
 		Assert.That( db.Assert( "A" ), Is.EqualTo( true ) );
+	}
 
-		// Update a value
-		db.Insert( "A.relationships.B.reputation!-99" );
-		Assert.That( db.Assert( "A.relationships.B.reputation.-99" ), Is.EqualTo( false ) );
+	/// <summary>
+	/// Test that create, retrieval, updating, and deletion (CRUD)
+	/// </summary>
+	[Test]
+	public void TestDeleteSentence()
+	{
+		var db = new RePraxisDatabase();
+
+		// Create values
+		db.Insert( "A.relationships.B.reputation!10" );
 
 		// Delete a value
 		db.Delete( "A.relationships.B.reputation" );
 		Assert.That( db.Assert( "A.relationships.B.reputation" ), Is.EqualTo( false ) );
+	}
+
+	/// <summary>
+	/// Test that create, retrieval, updating, and deletion (CRUD)
+	/// </summary>
+	[Test]
+	public void TestUpdateSentence()
+	{
+		var db = new RePraxisDatabase();
+
+		// Create values
+		db.Insert( "A.relationships.B.reputation!10" );
+
+		// Update a value
+		db.Insert( "A.relationships.B.reputation!-99" );
+		Assert.That( db.Assert( "A.relationships.B.reputation!-99" ), Is.EqualTo( true ) );
+		Assert.That( db.Assert( "A.relationships.B.reputation.-99" ), Is.EqualTo( false ) );
 	}
 
 	[Test]
@@ -39,8 +64,10 @@ public class Tests
 		db.Insert( "astrid.relationships.britt.tags.ex_lover" );
 		db.Insert( "astrid.relationships.lee.reputation!20" );
 		db.Insert( "astrid.relationships.lee.tags.friend" );
+		db.Insert( "britt.relationships.player.tags.spouse" );
 		db.Insert( "player.relationships.jordan.reputation!-20" );
 		db.Insert( "player.relationships.jordan.tags.enemy" );
+		db.Insert( "player.relationships.britt.tags.spouse" );
 
 		// Relational expression with a single variable
 		var r0 = new DBQuery()
@@ -75,7 +102,7 @@ public class Tests
 			.Run( db );
 
 		Assert.That( r2.Success, Is.EqualTo( true ) );
-		Assert.That( r2.Bindings.Length, Is.EqualTo( 1 ) );
+		Assert.That( r2.Bindings.Length, Is.EqualTo( 0 ) );
 
 		// Failing assertion without variables
 		var r3 = new DBQuery()
@@ -97,11 +124,85 @@ public class Tests
 		Assert.That( r4.Success, Is.EqualTo( true ) );
 		Assert.That( r4.Bindings.Length, Is.EqualTo( 1 ) );
 
-		// Check for negation
+		// Check that a sentence without variables is not true within the database
 		var r5 = new DBQuery()
 			.Where( "not player.relationships.jordan.reputation!30" )
 			.Run( db );
 
 		Assert.That( r5.Success, Is.EqualTo( true ) );
+
+		// Given that ?other is jordan, is the statement still not true?
+		var r6 = new DBQuery()
+			.Where( "not player.relationships.?other.reputation!30" )
+			.Run( db, new Dictionary<string, string>()
+			{
+				{"?other", "jordan"}
+			} );
+
+		Assert.That( r6.Success, Is.EqualTo( true ) );
+
+		// For all relationships astrid has with all ?others,
+		// no relationship has a reputation of 15
+		var r7 = new DBQuery()
+			.Where( "not astrid.relationships.?other.reputation!15" )
+			.Run( db );
+
+		Assert.That( r7.Success, Is.EqualTo( true ) );
+
+		// For all relationships astrid has with an ?other,
+		// filter for those where reputation is not 30
+		var r7_b = new DBQuery()
+			.Where( "astrid.relationships.?other.reputation!?rep" )
+			.Where( "neq ?rep 30" )
+			.Run( db );
+
+		Assert.That( r7_b.Success, Is.EqualTo( true ) );
+		Assert.That( r7_b.Bindings.Length, Is.EqualTo( 2 ) ); // britt and lee
+
+
+		// For all relationships astrid has with an ?other
+		// filter for those where reputation is not 30
+		var r8 = new DBQuery()
+			.Where( "astrid.relationships.?other" )
+			.Where( "not astrid.relationships.?other.reputation!30" )
+			.Run( db );
+
+		Assert.That( r8.Success, Is.EqualTo( true ) );
+		Assert.That( r8.Bindings.Length, Is.EqualTo( 2 ) ); // britt and lee
+
+		// For all relationships astrid has with an ?other
+		// filter for those where reputation from astrid to ?other is not 30
+		// and ?other does not have a relationship with a spouse tag
+		var r9 = new DBQuery()
+			.Where( "astrid.relationships.?other" )
+			.Where( "not astrid.relationships.?other.reputation!30" )
+			.Where( "not ?other.relationships.?others_spouse.tags.spouse" )
+			.Run( db );
+
+		Assert.That( r9.Success, Is.EqualTo( true ) );
+		Assert.That( r9.Bindings.Length, Is.EqualTo( 1 ) ); // lee
+
+		// For all relationships astrid has with an ?other
+		// filter for those where reputation from astrid to ?other is not 30.
+		// Also ensure that the player does not have a spouse
+		var r10 = new DBQuery()
+			.Where( "astrid.relationships.?other" )
+			.Where( "not astrid.relationships.?other.reputation!30" )
+			// below will fail because player has spouse
+			.Where( "not player.relationships.?x.tags.spouse" )
+			.Run( db );
+
+		Assert.That( r10.Success, Is.EqualTo( false ) );
+
+		// For all relationships astrid has with an ?other
+		// filter for those where reputation from astrid to ?other is not 30.
+		// Also ensure that the player does not have any friends
+		var r11 = new DBQuery()
+			.Where( "astrid.relationships.?other" )
+			.Where( "not astrid.relationships.?other.reputation!30" )
+			.Where( "not player.relationships.?x.tags.friend" )
+			.Run( db );
+
+		Assert.That( r11.Success, Is.EqualTo( true ) );
 	}
 }