+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on <https://common-changelog.org/>, and this project adheres mostly to Semantic Versioning. However, all releases before 1.0.0 have breaking changes between minor-version updates.
 
+## [1.2.0] - 2024-01-08
+
+### Added
+
+- Users can supply an array of bindings to queries
+- Add `QueryResult.ToPrettyString()` to simplify viewing query results
+- Update README to better explain query statements
+
+### Fixed
+
+- Not-statement inconsistent results from unclear semantics
+- Empty initial bindings causing query to fail
+
 ## [1.1.0] - 2024-01-05
 
 ### Added
@@ -27,3 +40,4 @@ _Initial release._
 
 [1.0.0]: https://github.com/ShiJbey/RePraxis/releases/tag/v1.0.0
 [1.1.0]: https://github.com/ShiJbey/RePraxis/releases/tag/v1.1.0
+[1.2.0]: https://github.com/ShiJbey/RePraxis/releases/tag/v1.2.0